+++ lib/homebrew-fork/Library/Homebrew/utils.rb
@@ -1,6 +1,5 @@
 require 'pathname'
 require 'exceptions'
-require 'os/mac'
 require 'hardware'
 require 'utils/popen'
